refactor: Add constants folder and Fix some files
