+++ app/constants/actionTypes.js
@@ -1,0 +1,2 @@
+export const INCREMENT_COUNTER = 'INCREMENT_COUNTER'
+export const DECREMENT_COUNTER = 'DECREMENT_COUNTER'
