+++ app/containers/HomePage.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import Home from '../components/Home';
+import Home from '../components/Home/Home';
 
 export default class HomePage extends Component {
   render() {
