+++ app/reducers/counter.js
@@ -1,10 +1,10 @@
-import { INCREMENT_COUNTER, DECREMENT_COUNTER } from '../actions/counter';
+import * as types from '../constants/actionTypes'
 
 export default function counter(state = 0, action) {
   switch (action.type) {
-    case INCREMENT_COUNTER:
+    case types.INCREMENT_COUNTER:
       return state + 1;
-    case DECREMENT_COUNTER:
+    case types.DECREMENT_COUNTER:
       return state - 1;
     default:
       return state;
