+++ package.json
@@ -79,10 +79,11 @@
     "fbjs-scripts": "^0.6.0",
     "jsdom": "^8.4.0",
     "json-loader": "^0.5.4",
+    "material-ui": "^0.14.4",
     "minimist": "^1.2.0",
     "mocha": "^2.4.5",
     "node-libs-browser": "^1.0.0",
-    "react-addons-test-utils": "^15.0.1",
+    "react-addons-test-utils": "^0.14.8",
     "redux-devtools": "^3.2.0",
     "redux-devtools-dock-monitor": "^1.1.1",
     "redux-devtools-log-monitor": "^1.0.11",
@@ -99,11 +100,12 @@
     "electron-debug": "^0.6.0",
     "font-awesome": "^4.6.1",
     "postcss": "^5.0.19",
-    "react": "^15.0.1",
-    "react-dom": "^15.0.1",
+    "react": "^0.14.8",
+    "react-dom": "^0.14.7",
     "react-redux": "^4.4.5",
     "react-router": "^2.2.4",
     "react-router-redux": "^4.0.2",
+
     "redux": "^3.4.0",
     "redux-thunk": "^2.0.1",
     "source-map-support": "^0.4.0"
