+++ test/e2e.js
@@ -3,8 +3,8 @@ import chromedriver from 'chromedriver';
 import webdriver from 'selenium-webdriver';
 import { expect } from 'chai';
 import electronPath from 'electron-prebuilt';
-import homeStyles from '../app/components/Home.css';
-import counterStyles from '../app/components/Counter.css';
+import homeStyles from '../app/components/Home/Home.css';
+import counterStyles from '../app/components/Counter/Counter.css';
 
 chromedriver.start(); // on port 9515
 process.on('exit', chromedriver.stop);
